responsible-ai-reporting-tool/wrapper/requirements/requirement.txt
uvicorn
pyyaml
fastapi==0.99.1
watchfiles
requests==2.25
python-multipart==0.0.5
email-validator
openpyxl
pymongo
python-dotenv
pdfkit
reportlab
PyPDF2
adversarial-robustness-toolbox
#lib/aicloudlibs-0.1.0-py3-none-any.whl
#lib/adversarial_robustness_toolbox-1.16.0-py3-none-any.whl
#Un-Comment Below 2 lines and comment above 2 lines to run locally
# ../lib/aicloudlibs-0.1.0-py3-none-any.whl  
#../lib/adversarial_robustness_toolbox-1.16.0-py3-none-any.whl
